EDU-395 refactor: model을 분리
